Backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.43.0
colorama==0.4.6
cron-descriptor==1.4.5
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.5
django-appconf==1.1.0
django-celery-beat==2.7.0
django-celery-email==3.0.0
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-js-asset==3.0.1
django-mptt==0.16.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
filelock==3.17.0
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
numpy==2.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
redis==5.2.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.1
wcwidth==0.2.13
whitenoise==6.9.0
